requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-ninja==1.3.0
django-ninja-extra==0.21.8
django-ninja-jwt==5.3.5
extensions==0.4
injector==0.22.0
mysqlclient==2.2.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
